move theme context
