+++ InvenTree/web/src/src/App.tsx
@@ -18,7 +18,7 @@ import { Profile } from './pages/Index/Profile/Profile';
 import Layout from './pages/layout';
 import { Login } from './pages/Login';
 import { Logout } from './pages/Logout';
-import { ThemeContext } from './ThemeContext';
+import { ThemeContext } from './context/ThemeContext';
 import { LanguageContext } from './translation';
 
 // Error tracking
